refactor: 번호 발급을 lotto 객체로 이동한다
