+++ src/main/java/lotto/LottoNumber.java
@@ -9,7 +9,6 @@ public class LottoNumber {
         this.value = value;
     }
 
-
     @Override
     public String toString() {
         return "LottoNumber{" +
