+++ src/main/java/lotto/Lottos.java
@@ -9,33 +9,17 @@ import static java.util.stream.Collectors.toList;
 import static java.util.stream.IntStream.rangeClosed;
 
 public class Lottos {
-    private static final int MIN = 1;
-    private static final int MAX = 45;
-    private static final List<Integer> NUMBERS = rangeClosed(MIN, MAX).boxed().collect(toList());
-
     private final List<Lotto> values;
 
     public Lottos(int quantity) {
         List<Lotto> lottos = new ArrayList<>();
 
         for (int i = 0; i < quantity; i++) {
-            Lotto lotto = new Lotto(pickLottoNumbers());
+            Lotto lotto = new Lotto();
             lottos.add(lotto);
         }
 
         this.values = lottos;
-    }
-
-    private LottoNumbers pickLottoNumbers() {
-        Collections.shuffle(NUMBERS);
-
-        List<LottoNumber> pickedNumbers = NUMBERS.stream()
-                .limit(6)
-                .sorted()
-                .map(LottoNumber::new)
-                .collect(toList());
-
-        return new LottoNumbers(pickedNumbers);
     }
 
     public List<Lotto> getValues() {
